+++ vpr/src/draw/draw.h
@@ -23,13 +23,25 @@ extern ezgl::application application;
 #endif /* NO_GRAPHICS */
 
 void update_screen(ScreenUpdatePriority priority, const char* msg, enum pic_type pic_on_screen_val, std::shared_ptr<SetupTimingInfo> timing_info);
+
+
 //Initializes the drawing locations.
 //FIXME: Currently broken if no rr-graph is loaded
 void init_draw_coords(float clb_width);
 
+
+/* Sets the static show_graphics and gr_automode variables to the    *
+ * desired values.  They control if graphics are enabled and, if so, *
+ * how often the user is prompted for input.                         */
 void init_graphics_state(bool show_graphics_val, int gr_automode_val, enum e_route_type route_type, bool save_graphics, std::string graphics_commands);
 
+
+/* Allocates the structures needed to draw the placement and routing.*/
 void alloc_draw_structs(const t_arch* arch);
+
+
+/* Free everything allocated by alloc_draw_structs. Called after close_graphics()
+ * in vpr_api.c. */
 void free_draw_structs();
 
 #ifndef NO_GRAPHICS
@@ -39,33 +51,44 @@ const ezgl::color DRIVEN_BY_IT_COLOR = ezgl::LIGHT_MEDIUM_BLUE;
 
 const float WIRE_DRAWING_WIDTH = 0.5;
 
+
+/* Find the edge between two rr nodes */
 t_edge_size find_edge(int prev_inode, int inode);
 
+
+/* Returns the track number of this routing resource node inode. */
 int get_track_num(int inode, const vtr::OffsetMatrix<int>& chanx_track, const vtr::OffsetMatrix<int>& chany_track);
+
 
 //Returns the drawing coordinates of the specified pin
 ezgl::point2d atom_pin_draw_coord(AtomPinId pin);
 
+
 //Returns the drawing coordinates of the specified tnode
 ezgl::point2d tnode_draw_coord(tatum::NodeId node);
 
-void annotate_draw_rr_node_costs(ClusterNetId net, int sink_rr_node);
-void clear_draw_rr_annotations();
 
-void draw_reset_blk_colors();
-void draw_reset_blk_color(ClusterBlockId blk_id);
-
+/* Converts a vtr Color to a ezgl Color. */
 ezgl::color to_ezgl_color(vtr::Color<float> color);
 
+
+/* This helper function determines whether a net has been highlighted. The highlighting
+ * could be caused by the user clicking on a routing resource, toggled, or
+ * fan-in/fan-out of a highlighted node. */
 bool draw_if_net_highlighted(ClusterNetId inet);
 std::vector<int> trace_routed_connection_rr_nodes(
     const ClusterNetId net_id,
     const int driver_pin,
     const int sink_pin);
+
+
+/* Helper function for trace_routed_connection_rr_nodes
+ * Adds the rr nodes linking rt_node to sink_rr_node to rr_nodes_on_path
+ * Returns true if rt_node is on the path. */
 bool trace_routed_connection_rr_nodes_recurr(const t_rt_node* rt_node,
                                              int sink_rr_node,
                                              std::vector<int>& rr_nodes_on_path);
-void draw_screen();
+
 
 /* This routine highlights the blocks affected in the latest move      *
  * It highlights the old and new locations of the moved blocks         *
@@ -84,6 +107,15 @@ void clear_colored_locations();
 // If the input loc is marked in colored_locations vector, the function will return true and the correspnding color is sent back in loc_color
 // otherwise, the function returns false (the location isn't among the highlighted locations)
 bool highlight_loc_with_specific_color(int x, int y, ezgl::color& loc_color);
+
+
+/* Because the list of possible block type colours is finite, we wrap around possible colours if there are more
+ * block types than colour choices. This ensures we support any number of types, although the colours may repeat.*/
+ezgl::color get_block_type_color(t_physical_tile_type_ptr type);
+
+
+/* Lightens a color's luminance [0, 1] by an aboslute 'amount' */
+ezgl::color lighten_color(ezgl::color color, float amount);
 
 #endif /* NO_GRAPHICS */
