+++ vpr/src/draw/draw_triangle.h
@@ -41,8 +41,27 @@
 #    include "place_macro.h"
 #    include "buttons.h"
 
+
+/*
+ * Draws a small triangle, at a position along a line from 'start' to 'end'.
+ * 'relative_position' [0., 1] defines the triangles position relative to 'start'.
+ * A 'relative_position' of 0. draws the triangle centered at 'start'.
+ * A 'relative_position' of 1. draws the triangle centered at 'end'.
+ * Fractional values draw the triangle along the line
+ */
 void draw_triangle_along_line(ezgl::renderer* g, ezgl::point2d start, ezgl::point2d end, float relative_position = 1., float arrow_size = DEFAULT_ARROW_SIZE);
+
+
+/* Draws a triangle with it's center at loc, and of length & width arrow_size,
+ * rotated such that it points in the direction of the directed line segment start -> end.
+ */
 void draw_triangle_along_line(ezgl::renderer* g, ezgl::point2d loc, ezgl::point2d start, ezgl::point2d end, float arrow_size = DEFAULT_ARROW_SIZE);
+
+/**
+ * Draws a triangle with it's center at (xend, yend), and of length & width arrow_size,
+ * rotated such that it points in the direction of the directed line segment (x1, y1) -> (x2, y2).
+ * Note parameter order.
+ */
 void draw_triangle_along_line(ezgl::renderer* g, float xend, float yend, float x1, float x2, float y1, float y2, float arrow_size = DEFAULT_ARROW_SIZE);
 
 #endif /* NO_GRAPHICS */
