+++ vpr/src/draw/intra_logic_block.h
@@ -90,16 +90,6 @@ struct t_selected_sub_block_info {
     bool is_in_selected_subtree(const t_pb_graph_node* test, const ClusterBlockId clb_index) const;
 };
 
-/* Enable/disable clb internals drawing. Internals drawing is enabled with a click of the
- * "Blk Internal" button. With each consecutive click of the button, a lower level in the
- * pb_graph will be shown for every clb. When the number of clicks on the button exceeds
- * the maximum level of sub-blocks that exists in the pb_graph, internals drawing
- * will be disabled.
- */
-#    ifndef NO_GRAPHICS
-void toggle_blk_internal(GtkWidget* /*widget*/, gint /*response_id*/, gpointer data);
-
-#    endif /* NO_GRAPHICS */
 /* This function pre-allocates space to store bounding boxes for all sub-blocks. Each
  * sub-block is identified by its descriptor_type and a unique pin ID in the type.
  */
